feat: add admin notifications feature
- Import new admin notifications CSS and JS modules.
- Create a notifications controller to manage fetching and rendering notifications.
- Implement badge label formatting based on notification count.
- Add loading states and error handling for notifications.
- Refactor existing admin shortcuts code to integrate the new notifications system.
- Update internationalization data for notification labels in both Polish and English.
- Introduce shared date formatting utility for consistent date display.
