+++ public/assets/css/styles.css
@@ -10,6 +10,7 @@
 @import "./components/forms.css";
 @import "./components/flash.css";
 @import "./components/admin-shortcuts.css";
+@import "./components/admin-notifications.css";
 @import "./components/admin-preloader.css";
 @import "./components/editor-select.css";
 @import "./components/dropdown-menu.css";
