+++ public/assets/js/modules/i18n-data.js
@@ -154,6 +154,10 @@ export const i18n = {
     admin_shortcut_notifications_mark_unread: "Oznacz jako nieprzeczytane",
     admin_shortcut_notifications_delete: "Usuń powiadomienie",
     admin_shortcut_notifications_delete_all: "Usuń wszystkie powiadomienia",
+    admin_shortcut_notifications_badge_one: "{{count}} powiadomienie",
+    admin_shortcut_notifications_badge_few: "{{count}} powiadomienia",
+    admin_shortcut_notifications_badge_many: "{{count}} powiadomień",
+    admin_shortcut_notifications_badge_other: "{{count}} powiadomienia",
     admin_shortcut_blog_settings: "Ustawienia bloga",
     admin_shortcut_import_export_title: "Importy & Eksporty",
     admin_shortcut_queue_status: "Stan kolejek",
@@ -931,6 +935,8 @@ export const i18n = {
     admin_shortcut_notifications_mark_unread: "Mark as unread",
     admin_shortcut_notifications_delete: "Delete notification",
     admin_shortcut_notifications_delete_all: "Delete all notifications",
+    admin_shortcut_notifications_badge_one: "{{count}} notification",
+    admin_shortcut_notifications_badge_other: "{{count}} notifications",
     admin_shortcut_blog_settings: "Blog settings",
     admin_shortcut_queue_status: "Queue status",
     admin_shortcut_import_export_title: "Imports & Exports",
